# perf(render): Don't set context.no_cache